requirements.txt
python-telegram-bot==20.7
pytz==2023.4
openai==0.28.0
